refactor(tests): extract shared test helpers into test_helpers.rs
Eliminates ~560 lines of duplicated MockOlapClient, create_test_project,
create_test_table, make_test_dict, make_infra_map_with_dict code that was
independently maintained in both infra_reality_checker.rs and plan.rs.
